Added support for testing with Vagrant.
Vagrantfile and supporting files for the successful installation
of the chef server and uploading of recipes using spiceweasel/berkshelf.

README.md updated with howto use instructions.  Tested with just the
os-base role, but plan is to have a vm definition for the allinone
role when all the cookbooks are ready.

Change-Id: I9b93c4054677013966aececa3093711d0c91bcbd